refactor: move prepare command to dataloss sub command
